feat(session): add vmux_session with the session components
The first crate that deliberately links no bevy_cef, so the headless server
has something to own session state through. Holds AgentSession,
AgentMessages, AgentConversationTitle, AgentApprovalPolicy, PromptQueue,
AgentVariant and AcpSession.

Nothing needed moving except AgentKind, done separately: Message,
AgentAttachment and ProcessId were already in vmux_wire, reached through
one-line re-exports in vmux_service and vmux_core, so the imports swap
without changing type identity.

AcpSession's own closure was always clean; only the file it sat in pulled
CEF, vmux_layout and vmux_terminal. The rest of that file stays.

components.rs survives as a shim because eighteen files import through it.
variant.rs is deleted outright, having exactly one re-export site.
