+++ crates/page/vmux_agent/Cargo.toml
@@ -44,6 +44,7 @@ vmux_layout = { path = "../vmux_layout" }
 vmux_mcp = { path = "../../host/vmux_mcp" }
 vmux_remote = { path = "../../host/vmux_remote" }
 vmux_service = { path = "../../host/vmux_service" }
+vmux_session = { path = "../../vmux_session" }
 vmux_setting = { path = "../vmux_setting" }
 vmux_space = { path = "../vmux_space" }
 vmux_terminal = { path = "../vmux_terminal" }
